08/05/2026 Bug Patch, UE update and improving clarity 	modified:   Collagens.py 	modified:   __pycache__/Collagens.cpython-313.pyc 	new file:   static/images/Wiki_Collagen_Logo.png 	deleted:    static/images/logo.jpg 	modified:   templates/citation.html 	modified:   templates/collagens.html 	modified:   templates/contact.html 	modified:   templates/documentation.html 	modified:   templates/errors.html 	modified:   templates/home.html 	modified:   templates/message_sent.html Patching sessions/.gitkeep deletion error in "Collagens.py" Patching polyglycine inclusion issue in "Collagens.py" Adding Wiki Collagen Logo for Favicon Altering Jordi Bella -> Dr. Jordi Bella Adding citation and GitHub link to "Citation" webpage

+++ Collagens.py
@@ -188,4 +188,5 @@ def cleanup_files(folder="sessions", age_limit=3600):
         path = os.path.join(folder, session)
         age = now - os.path.getmtime(path)
         if age > age_limit:
-            shutil.rmtree(path)
+            if path != "sessions\\.gitkeep":
+                shutil.rmtree(path)
